2007-06-04  Thomas Wagner <tom68@users.sourceforge.net>
        * SFEmpd.spec: bump to 0.13.0, re- --enable-flac --enable-oggflac
                       mpd now compiles with newer flac versions
                       (new feature) --enable-shout for streaming:t
                       "the net is the music"
                       if SFEavahi is present, mpd resources will
                       be announced with zeroconf/avahi/mDNS broadcasts
                       patch3: make id3_charset in mpdconf.example
                       default to UTF-8

        * SFEavahi.spec: (new) implemetation of service discovery in
                       local LAN (see Rendezvous/Bonjour/Zeroconf and ...
                       nwam in the future)
        * SFElibdaemon.spec: (new) library for building daemons (for avahi)

        * SFEgmpc.spec: (new feature) AVAHI zeroconf support
                        (new feature) icecast/shout stream to local client (needs: ogg123)
                        bump to 0.15.0, set compiler to gcc
        * SFEgmpc-(lyrics|serverstats|coveramazon|autoplaylist|qosd|stopbutton|lastfm|magnatune|osd|mdcover).spec
                        bump to 0.15.0, set compiler to gcc
        * SFElibmpd.spec: bump to 0.14.0
                        temporary patch libmpd-01-libmpdclient-include-limits.h.diff

        * SFExosd.spec: set compiler to gcc

        * SFEgtodo2.spec: (new) GNOME todo list with priorities and due date

        * SFEcurl.spec: bump to 7.16.2, --disable-static

        * SFExchat.spec: bump to 2.8.2
                         /usr/bin/msgfmt errors, use /opt/sfw/bin/msgfmt
                         reworked patch for 2.8.2

        * SFEicecast.spec: (new) stream musik over http to the LAN/WAN network
        * SFElibshout.spec: (new) Support Library for icecast audio streaming

        * SFEpulseaudio.spec: (new) first shot.
        * SFElibsamplerate.spec: (new) samplerate conversion for pulseaudio
        * SFEpavumeter.spec: (new) volume meter

        * SFEfont-terminus.spec: (new) nice readably font

        * SFElynx.spec: bump to 2.8.6, new download-URL since not in /current/-diry




git-svn-id: https://pkgbuild.svn.sourceforge.net/svnroot/pkgbuild/spec-files-extra/trunk@300 3354ce84-6216-0410-ad3d-afc617165c18

+++ SFExosd.spec
@@ -2,6 +2,8 @@
 # spec file for package SFExosd
 #
 
+# Note: builds only with  gcc !
+
 %include Solaris.inc
 Name:                    SFExosd
 Summary:                 xosd - simple library to display shaped text on X Display
@@ -38,7 +40,8 @@ Requires:                %{name}
 
 %build
 
-./configure --prefix=%{_prefix}  \
+
+CC=/usr/sfw/bin/gcc CXX=/usr/sfw/bin/g++ ./configure --prefix=%{_prefix}  \
             --mandir=%{_mandir}   \
             --disable-static
 
@@ -50,19 +53,9 @@ rm -rf $RPM_BUILD_ROOT
 make install DESTDIR=$RPM_BUILD_ROOT
 
 #empty directories (xmms):
-rm -r $RPM_BUILD_ROOT/opt
+test -d $RPM_BUILD_ROOT/opt && rm -r $RPM_BUILD_ROOT/opt
 
 rm -r $RPM_BUILD_ROOT/%{_libdir}/*la
-
-%if %build_l10n
-#TODO check if needed  # Rename pl_PL dir to pl as pl_PL is a symlink to pl and causing installation
-#TODO check if needed  # problems as a dir.
-#TODO check if needed  cd $RPM_BUILD_ROOT%{_datadir}/locale
-#TODO check if needed  mv pl_PL pl
-%else
-# REMOVE l10n FILES
-rm -rf $RPM_BUILD_ROOT%{_datadir}/locale
-%endif
 
 %clean
 rm -rf $RPM_BUILD_ROOT
@@ -100,5 +93,9 @@ rm -rf $RPM_BUILD_ROOT
 
 
 %changelog
+* Thu May 31 2007  - Thomas Wagner
+- force compiler to gcc.
+- conditional rm -r $RPM_BUILD_ROOT/opt (really needed?)
+- builds only with gcc
 * Thu Apr 07 2007  - Thomas Wagner
 - Initial spec
